Got link to open in new page - Not all places work
Google seems to have revoked ways to write direct links to its map. 
It can show the direct Lat and Lng but not the name of the places. So current one is the best way to write link. 
This one even doesn't work for every link. Some of them work.